feat: drop collection
