--- src/polodb_core/backend/file/mod.rs
@@ -1,12 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- */
-mod frame_header;
-mod transaction_state;
-mod journal_manager;
-mod file_backend;
-mod pagecache;
-
-pub(crate) use file_backend::FileBackend;
