--- src/polodb_core/backend/indexeddb/mod.rs
@@ -1,8 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- */
-mod indexeddb_backend;
-
-pub(crate) use indexeddb_backend::IndexedDbBackend;
