--- src/polodb_core/backend/memory/mod.rs
@@ -1,9 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- */
-mod memory_backend;
-mod db_snapshot;
-
-pub(crate) use memory_backend::MemoryBackend;
