+++ src/polodb_core/lsm/lsm_session.rs
@@ -1,9 +1,7 @@
 use std::io::Write;
 use byteorder::WriteBytesExt;
-use libc::printf;
 use crate::lsm::mem_table::MemTable;
 use crate::{DbErr, DbResult, TransactionType};
-use crate::lsm::KvCursor;
 use crate::lsm::lsm_backend::lsm_log::format;
 use crate::lsm::multi_cursor::MultiCursor;
 use crate::utils::vli;
@@ -112,14 +110,15 @@ impl LsmSession {
             LsmSession::put_log(log_buffer, key, value)?;
         }
 
-        let new_tree_opt = cursor.update_current(value);
+        let new_tree_opt = cursor.update_current(value)?;
         if let Some((new_tree, legacy_value_opt)) = new_tree_opt {
             self.mem_table.update_root(new_tree);
 
             if let Some(legacy_value) = legacy_value_opt {
                 *self.mem_table.store_bytes_mut() -= legacy_value.len();
-                result = true;
             }
+
+            result = true;
 
             *self.mem_table.store_bytes_mut() += value.len();
         }
@@ -139,7 +138,7 @@ impl LsmSession {
             LsmSession::delete_log(log_buffer, key)?;
         }
 
-        let new_tree_opt = cursor.delete_current();
+        let new_tree_opt = cursor.delete_current()?;
         if let Some((new_tree, legacy_value_opt)) = new_tree_opt {
             self.mem_table.update_root(new_tree);
 
@@ -147,8 +146,9 @@ impl LsmSession {
                 // The "key" and "mark" still needs space
                 // only substract the space of value here
                 *self.mem_table.store_bytes_mut() -= legacy_value.len();
-                result = true;
             }
+
+            result = true;
         }
 
         Ok(result)
